+++ HavenBridge.Api/Models/DonationAllocation.cs
@@ -21,7 +21,7 @@ public class DonationAllocation
     public string? ProgramArea { get; set; }
 
     [Column("amount_allocated")]
-    public float AmountAllocated { get; set; }
+    public double AmountAllocated { get; set; }
 
     [Column("allocation_date")]
     public DateOnly? AllocationDate { get; set; }
