+++ HavenBridge.Api/Models/InKindDonationItem.cs
@@ -29,7 +29,7 @@ public class InKindDonationItem
     public string? UnitOfMeasure { get; set; }
 
     [Column("estimated_unit_value")]
-    public float? EstimatedUnitValue { get; set; }
+    public double? EstimatedUnitValue { get; set; }
 
     [Column("intended_use")]
     [MaxLength(300)]
